+++ src/components/Team/Match/MatchSchedule.tsx
@@ -1,11 +1,14 @@
 import CommonButton from '../CommonButton/CommonButton';
 import MatchCard from './MatchCard';
-import { TeamName, teamDisplayNameMap } from '../../../utils/team-name-map';
-import { teamEmblemMap } from '../../../utils/team-emblem-map';
+import {
+  TeamName,
+  teamDisplayNameMap,
+} from '../../../utils/team/team-name-map';
+import { teamEmblemMap } from '../../../utils/team/team-emblem-map';
 import { useEffect, useState } from 'react';
 import { getTeamSchedule } from '../../../services/teamService';
 import { useQuery } from '@tanstack/react-query';
-import { teamLinks } from '../../../utils/team-links';
+import { teamLinks } from '../../../utils/team/team-links';
 
 type Match = {
   date: string;
